move css, make firendlist input css
