+++ frontend/src/css/FriendList.css
@@ -1,0 +1,7 @@
+.friendInput {
+  width: 768px;
+  padding: 1rem;
+  border-radius: 10px;
+  box-sizing: border-box;
+  font-weight: bold;
+}
