+++ frontend/src/css/index.css
@@ -1,10 +1,11 @@
-@import url('https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css');
+@import url("https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css");
 
 * {
   font-family: Pretendard, sans-serif;
 }
 
-body, html {
+body,
+html {
   height: 100%;
 }
 
@@ -20,13 +21,14 @@ button {
   padding: 1rem;
   border-radius: 10px;
   border-style: none;
-  background-color: #53B5EC;
+  background-color: #53b5ec;
   color: white;
   font-weight: bold;
+  margin-top: 10px;
 }
 
 code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
     monospace;
 }
 
@@ -34,6 +36,7 @@ h1 {
   font-size: 48px;
 }
 
-p, button {
+p,
+button {
   font-size: 18px;
 }
